fix: Rename generator exports, introduce integ tests package
As a pre-requisite for more generator tooling the generator export names
are normalized to `generate`. This allows the `jsugen` command to
dynamically require the function from configured generator packages
(both from core and user-land packages).

The integration tests are moved to a dedicated and private
`jsugen-integration-tests` package. The most prominent reason is that
testing a future `jsugen` command will require dynamic imports of
`generate` functions from installed modules. The separate package allows
specifying all core generators as a dependency so that they will be
available when the integration tests execute.

Additionally integration tests will also assert on file system
operations, which is easer in less dangerous in dediated sub-package.